Add unit tests for gencoupling (#89)
* Add validation to ensure max < min distance and valid bond lengths

* Add doc string for undocumented parameter of gencoupling

* Add unit tests for gencoupling

* Change to use default coupling struct and verify_val

* Add tests for forcenoSym, fid and atom dist < dMin

* Validate that all distances must be positive and add test

* Fix bug if no bond length > maxSym (now throw warning) and add test

* Add more detailed test for maxSym option when multiple bonds

* Fix typo in warniong string

* Add test that throws SymProblem error

Not sure if this is functioning as expected.
Bonds along a and b are sym. equiv. accroding to P 4 group but the lengths differ by d=0.5 Ang.
Here an error is thrown when d < tol but surely that is then consistent (i.e. tol is tolerance in bond lengths considered equiv.).
Also I think it should actually be tolDist (tolD variable in code) not tol used in this comparison.

* Use tolDist for equivalence if spacegrp and change tol to tolMaxDist

* Update test with P2 sym to check inequiv. bonds with same length

Check that when symetry of lattice is higher than spacegorup provided the bonds of same length are correctly indexed as inequivilent based on spacegroup symmetry operations.

* Throw error if tolerances larger than dMax (and added tests)

* Update changelog with breaking change to API (param name changed)

* Add test for two atoms (mag/non-mag) in unit cell

* Update changelog with added validation

* Add line breaks to long error msg in gencoupling